feat: CollaborationService 구현 및 패키지 구조 변경
