+++ backend/api/src/main/java/com/yat2/episode/collaboration/config/WebSocketConfig.java
@@ -1,4 +1,4 @@
-package com.yat2.episode.collaboration;
+package com.yat2.episode.collaboration.config;
 
 
 import lombok.RequiredArgsConstructor;
@@ -6,6 +6,9 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.web.socket.config.annotation.EnableWebSocket;
 import org.springframework.web.socket.config.annotation.WebSocketConfigurer;
 import org.springframework.web.socket.config.annotation.WebSocketHandlerRegistry;
+
+import com.yat2.episode.collaboration.ws.HandshakeInterceptor;
+import com.yat2.episode.collaboration.ws.RelayHandler;
 
 @Configuration
 @EnableWebSocket
