+++ backend/api/src/main/java/com/yat2/episode/collaboration/ws/RelayHandler.java
@@ -1,4 +1,4 @@
-package com.yat2.episode.collaboration;
+package com.yat2.episode.collaboration.ws;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
@@ -7,10 +7,12 @@ import org.springframework.web.socket.CloseStatus;
 import org.springframework.web.socket.WebSocketSession;
 import org.springframework.web.socket.handler.AbstractWebSocketHandler;
 
+import com.yat2.episode.collaboration.CollaborationService;
+
 @RequiredArgsConstructor
 @Component
 public class RelayHandler extends AbstractWebSocketHandler {
-    private final MessageService messageService;
+    private final CollaborationService collaborationService;
 
     @Override
     public void afterConnectionEstablished(WebSocketSession session) {
